requirements.txt
asgiref==3.8.1
boto3
channels==4.0.0
channels-redis
Django==4.2
django-cors-headers
django-extensions
django-ratelimit==4.1.0
django-storages
django-widget-tweaks
djangorestframework
djangorestframework-simplejwt
mysqlclient
Pillow
python-dotenv
python-magic-bin
razorpay==1.4.1
redis
whitenoise
yfinance
